Given a function f, known at evenly spaced samples y_j = f(a + jh), this function constructs an interpolant using compactly supported cubic b splines. The advantage of using splines of compact support over traditional cubic splines is that compact support makes the splines well-conditioned.
The interpolant is constructed in O(N) time and can be evaluated in constant time.
Its error is O(h^4), and obeys the interpolating condition s(x_j) = f(x_j) for all samples.
In addition, f' can be estimated from s', albeit with lower accuracy.

This routine is cppcheck clean, and is clean under AddressSanitizer and MemorySanitizer.