Addition of the barycentric rational interpolant.
This commit can be regarded as a "friend" to the cubic b spline, filling out the feature set.
Whereas the cubic b spline interpolates uniformly spaced data, the barycentric rational interpolant can interpolate arbitrarily spaced data.
The price that is paid is that evaluation of the interpolant is O(N), rather than O(1) for the cubic b spline interpolant.

The routine supports essentially arbitrary approximation order d; with error of O(h^d+1).
It is cppcheck-clean, run under AddressSanitizer and UndefinedBehaviorSanitizer, and works with boost::multiprecision.

References:

1) Floater, M.S. & Hormann, K. Numer. Math. (2007) 107: 315. doi:10.1007/s00211-007-0093-y

2) Numerical Recipes, Section 3.4.1