Tanh-sinh quadrature
This class performs tanh-sinh quadrature on holomorphic functions beloning to Hardy spaces.
Although this seems like a dire restriction, in fact there are many functions which obey fall into this class
as is shown in the unit tests.

The class compiles under gcc and clang, runs cleanly under AddressSanitizer and MemorySanitizer,
and is cppcheck-clean.

Unfortunately, variable remappings require the use of lambdas, and as such this cannot be compiled without C++11.

Finally, a large collection of ascissas and weights is used in the class.
These can be updated by the 'generate_constants()' function in the unit tests.
Performance tests show that on a 3GHz intel CPU, a double precision linear function can be integrated in 150 ns,
a quadratic in 183 ns. Full benchmarks for many different precisions have been performed and show the expected scaling.