Complex Step Differentiation and Optimal stepsize finite difference differentiation
This commit adds the complex step differentiation, which is an unconditionally stable algorithm for computing derivatives.
However, the function must be holomorphic and take on real values at real arguments.
When this requirement is satisfied, this algorithm computes the derivative to the machine precision in a single function evaluation.

The requirement that the function be holomorphic is a strong one; for instance Daubechies wavelets cannot be differentiated this way
and considerable coding effort must be put into writing functions which can be evaluated at both real and complex arguments.
Hence finite-difference formulas using optimal stepsizes are provided which only require local C^{k+1}-differentiability for k-order accuracy.
It is well-known that finite-differencing is unstable, but nonetheless,
good results are often obtained in spite of this.
Error estimates are provided to help cope with instabilities.