Fixed type mismatch in exp_sinh quadrature
Fixed bug in quadrature::exp_sinh, which caused it to fail to compile
when used with a complex valued integrand over a non-native range. This
was due to the lambdas perfoming the change of variables to shift onto
the native range requiring the return type to be Real, rather than using
the type infered from the integrand.