Policy support for Arithmetic-Geometric Mean
As mentioned in #374, agm currently hangs when invoked with the
real_concept type. This is because the function:

    std::numeric_limits<real_concept>::epsilon()

Returns zero, thus putting agm into an infinite loop. The current
patch adapts the agm code to use floating-point policies, in particular
`policies::get_epsilon<Real, Policy>()`, and introduces a second
function signature to accept a Policy argument. The new signature is
added to the documentation, and the tests are updated to use the
Policy'ed version of agm. Tests are also added to cover the real_concept
type to prove that the original AGM hang no longer occurs.

Two other small changes are introduced here: first, a trivial
implementation of isnan() for the real_concept type (necessary to get
the new tests to compile and run), and second a small typo fix in one of
the AGM test messages.