Expand the document of Catmull-Rom spline
It is not clear from the current documentation how the curve parameter is relates to how to interpolate between any two of the input points, which is, at the end of the day, one reason for using spline interpolation. Here a blurb is added to the doc, following https://github.com/boostorg/math/issues/211.